pyproject.toml
[project]
name = "koffe"
version = "0.1.0"
description = "Specialty coffee browser — scrapes roasters, stores data, serves a filterable UI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.40",
    "selectolax>=0.3",
    "httpx>=0.26",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "apscheduler>=3.10",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
